requirements.txt
# AutoDataPreprocessor Requirements

# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0

# API client
requests>=2.25.0
groq>=0.4.0  # For Groq API
python-dotenv>=0.19.0

# Data validation and processing
jsonschema>=4.0.0
tenacity>=8.0.0  # For retry logic

# Development and testing
pytest>=7.0.0
black>=22.0
flake8>=4.0.0
mypy>=0.900